requirements.txt
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.9.1
arrow==1.2.2
asgiref==3.5.0
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
blinker==1.4
bs4==0.0.1
CacheControl==0.12.11
cached-property==1.5.1
cachy==0.3.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
chevron==0.14.0
cleo==0.8.1
click==8.1.2
clikit==0.6.2
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cookiecutter==1.7.3
crashtest==0.3.1
cryptography==2.8
cssselect==1.1.0
cycler==0.11.0
dbus-python==1.2.16
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
easydict==1.9
entrypoints==0.3
fake-useragent==0.1.11
fastapi==0.73.0
filelock==3.6.0
fonttools==4.29.1
greenlet==1.1.2
h11==0.12.0
html5lib==1.1
httpcore==0.14.7
httplib2==0.14.0
httptools==0.4.0
httpx==0.22.0
hyperlink==19.0.0
idna==2.8
imageio==2.16.0
importlib-metadata==4.11.3
incremental==16.10.1
isc==2.0
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.1
jinja2-time==0.2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.6.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
more-itertools==4.2.0
msgpack==1.0.3
multidict==6.0.2
nb-cli==0.6.6
netifaces==0.10.4
nonebot-adapter-onebot==2.0.0b1
nonebot-plugin-apscheduler==0.1.2
nonebot-plugin-gocqhttp==0.5.2
nonebot2==2.0.0b2
numpy==1.22.1
oauthlib==3.1.0
opencv-python==4.5.5.62
packaging==20.9
pandas==1.4.1
parse==1.19.0
pastel==0.2.1
pexpect==4.8.0
Pillow==9.0.0
pkginfo==1.8.2
platformdirs==2.5.2
ply==3.11
poetry==1.1.13
poetry-core==1.0.8
poyo==0.5.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycurl==7.43.0.2
pydantic==1.9.0
pyee==8.2.2
pyfiglet==0.8.post1
PyGObject==3.36.0
pygtrie==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pylev==1.4.0
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pypinyin==0.46.0
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-slugify==6.1.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
requests-html==0.10.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
scipy==1.8.0
SecretStorage==3.3.2
service-identity==18.1.0
shellingham==1.4.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
ssh-import-id==5.10
starlette==0.17.1
systemd-python==234
text-unidecode==1.3
texttable==1.6.2
tomlkit==0.9.2
torch==1.10.1+cpu
torchaudio==0.10.1+cpu
torchvision==0.11.2+cpu
tqdm==4.64.0
Twisted==18.9.0
typing-extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
w3lib==1.22.0
wadllib==1.3.3
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
websockets==10.3
wordcloud==1.8.1
yarl==1.7.2
zipp==1.0.0
zope.interface==4.7.1